% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{zincSSD}
\alias{zincSSD}
\title{Function that fits ssd to burrIII distribution
Plot end result}
\usage{
zincSSD(input, hcv = NULL)
}
\description{
Function that fits ssd to burrIII distribution
Plot end result
}
